requirements.txt
streamlit
streamlit_chat
langchain
sentence_transformers
google-cloud-storage
unstructured
pinecone-client
google-cloud-aiplatform